move external/project dependency visitors to :modulecheck-psi
